+++ hutool-extra/src/main/java/cn/hutool/extra/template/engine/freemarker/package-info.java
@@ -1,7 +1,7 @@
 /**
- * Freemarker实现
- * 
- * @author looly
+ * Freemarker实现<br>
+ * 见：https://freemarker.apache.org/
  *
+ * @author looly
  */
 package cn.hutool.extra.template.engine.freemarker;